FAT
Write a java program application to validate the username and password of a user.
Read the username and password as strings, if the username is valid and password is correct, then display, "Welcome, Username". 
Display "Invalid Username", if the username is invalid. Display "Wrong Password", if the password is wrong.
Valid Username must be:
1) Minimum - 8 characters, maximum - 32 characters
2) Can contain alphabets (a-z, A-Z)  and digits(0-9).
3) The only special character allowed is underscore _.
4) Username must start with an alphabet.
The allowed passwords are
admin123
VITadmin! 
StudScope153*
Validate the password, only if the username is valid.
Sample I/O:
9483user
admin123
Invalid Username

Raj123!
admin123
Invalid Username

Dev1991_home
VITadmin!
Welcome, Dev1991_home