PAT
Kids Play school is planning for a slow walk race for 'n' children. The kid who touches the finishing line first gets 0 chocolate.
The second kid gets 1 chocolate. The numer of chocolates the third kid gets is the sum of chocolates given to first and second kid. 
Fourth kid gets sum of chocolates of second and third kids and it goes on. Print the number of chocolates got by each kid. 
Sample Input:
5 - Number of Children - Integer
Sample Output:
0 1 1 2 3