PAT
Read an array of 'n' integers, check whether each integer is a prime number or not. If any given number is 0 or 1, 
display "It is neither prime nor composite". Also print the numbers which are perfect numbers in the given list. 
Also print the sum of even numbers and odd numbers from the given list.
Perfect Number - a number that is equal to the sum of its proper divisors.
Prime Number - a number that can be divided exactly only by itself and 1
Sample Input:
8
1 23 6 9 41 0 71 74
Sample Output
1 is neither prime nor composite
23 is a prime number
6 is not a prime number
9 is not a prime number
41 is a prime number
0 is neither prime nor composite
71 is a prime number
74 is not a prime number
Perfect Numbers:
6
Even Sum:80
Odd Sum:145