PAT
Given an Integer n, check whether it is an Armstrong number. Also check whether it is a perfect number. 
Also print the sum of even digits and sum of odd digits in the given number.
(Armstrong number is a number that is equal to the sum of cubes of its digits. Perfect number is an integer that is equal to the sum of its proper divisors)
Define the following functions with the given signatures:
1. boolean isArmstrong(int) 
2. boolean isPerfect(int)
3. void calcSum(int)
Sample Input:
I53 - Integer n
Sample output:
It is an Armstrong Number 
It is not a Perfect Number
Even Sum: 0
Odd Sum: 9